Add horizontal rules to groups question with its how verified
We need to ask a thing and how it was verified. The how verified
questions do not include then thing in its title so we rely on proximity
for context. I think by including horizontal rules between the question
groups they will be easier to use.

For non-sighted users we have a h2 for the thing and a h3 for the how
verified so that they can navigate by h2s to quickly jump between
sections.