:speech_balloon: hint explicitly says carry on then return here later
Users were getting hung up on the fact that they had to to do checks and
some did not want to proceed without having the results. We added some
copy to say they could do the checks and return later but it did not
explicitly state that they could complete later sections of the report
then return to answer about the checks. It now does this.